# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Real Estate Investment API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /user/profile:
    get:
      operationId: getUserProfile
      tags: [user]
      summary: Get user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserFull"
    patch:
      operationId: updateUserProfile
      tags: [user]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserFull"
  /user/avatar:
    patch:
      operationId: updateAvatar
      tags: [user]
      summary: Update user avatar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAvatarBody"
      responses:
        "200":
          description: Avatar updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
  /user/password:
    patch:
      operationId: changePassword
      tags: [user]
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordBody"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
...
  /user/change-pin:
    post:
      operationId: changeTransactionPin
      tags: [user]
      summary: Change transaction PIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePinBody"
      responses:
        "200":
          description: PIN updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChangePinResponse"
  /user/transfer:
    post:
      operationId: userTransfer
      tags: [user]
      summary: Transfer balance to another user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserTransferBody"
      responses:
        "200":
          description: Transfer successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserTransferResponse"
...
components:
  schemas:
...
    UpdateAvatarBody:
      type: object
      required: [avatarUrl]
      properties:
        avatarUrl:
          type: string
    SuccessMessage:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
...
    UserFull:
      type: object
      required:
        - id
        - firstName
        - middleName
        - surname
        - phone
        - whatsappNumber
        - username
        - email
        - gender
        - avatar
        - homeAddress
        - bankName
        - accountNumber
        - accountHolderName
        - zipCode
        - referralCode
        - referredBy
        - position
        - level
        - role
        - isActive
        - balance
        - securityDeposit
      properties:
        id:
          type: integer
        firstName:
          type: string
        middleName:
          type: string
        surname:
          type: string
        phone:
          type: string
        whatsappNumber:
          type: string
        username:
          type: string
        email:
          type: string
        gender:
          type: string
        avatar:
          type: string
        activatedLevels:
          type: array
          items:
            type: string
        homeAddress:
          type: string
        bankName:
          type: string
        accountNumber:
          type: string
        accountHolderName:
          type: string
        zipCode:
          type: string
        referralCode:
          type: string
        referredBy:
          type: string
        position:
          type: string
        level:
          type: string
        role:
          type: string
        isActive:
          type: boolean
        withdrawalLocked:
          type: boolean
        balance:
          type: number
        securityDeposit:
          type: number
        weeklyTransferUsed:
          type: number
        weeklyTransferLimit:
          type: number
    UpdateProfileBody:
      type: object
      properties:
        firstName:
          type: string
        middleName:
          type: string
        surname:
          type: string
        whatsappNumber:
          type: string
        homeAddress:
          type: string
        bankName:
          type: string
        accountNumber:
          type: string
        accountHolderName:
          type: string
        zipCode:
          type: string
    ChangePasswordBody:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
...
    UserTransferBody:
      type: object
      required: [recipientUsername, amount, transactionPin]
      properties:
        recipientUsername:
          type: string
        amount:
          type: number
        transactionPin:
          type: string
    UserTransferResponse:
      type: object
      required: [success, message, newBalance]
      properties:
        success:
          type: boolean
        message:
          type: string
        newBalance:
          type: number
...
    ChangePinBody:
      type: object
      required: [newPin]
      properties:
        currentPin:
          type: string
        newPin:
          type: string
    ChangePinResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
...